Add command gozmqfix.
This is a command that will update code when there are backwards
incompatible changes in the API of the "github.com/alecthomas/gozmq"
package.  This commit brings in the skeleton with no specific API
updates yet included.

The code is copied from the Go project at version 1.0.3 with its
licensing intact.  Based on a little research into licensing, I've added
a copyright line near the top of the LICENSE file.  I am not a lawyer.

Install with `go install github.com/alecthomas/gozmq/gozmqfix`.  Usage
is essentially the same as for `go fix`.